+++ tools/unreal_fix/0.38.4/original/wldr_isd.cpp
@@ -20,8 +20,8 @@ void FDD::format_isd()
          for(unsigned s = 0; s < 5; s++)
          {
             unsigned n = sn[s];
-            t.hdr[s].n = n; t.hdr[s].l = 3;
-            t.hdr[s].c = c; t.hdr[s].s = 0;
+            t.hdr[s].n = u8(n); t.hdr[s].l = 3;
+            t.hdr[s].c = u8(c); t.hdr[s].s = 0;
             t.hdr[s].c1 = t.hdr[s].c2 = 0;
             t.hdr[s].data = (unsigned char*)1;
          }
